msm8953-common: Motoactions: Convert over to INT
Fixes issues with not working and showing that its covered
and its not

+++ MotoActions/src/com/moto/actions/actions/ChopChopSensor.java
@@ -75,7 +75,7 @@ public class ChopChopSensor implements SensorEventListener, UpdatedStateNotifier
     private SensorEventListener mProxListener = new SensorEventListener() {
         @Override
         public synchronized void onSensorChanged(SensorEvent event) {
-            mProxIsCovered = event.values[0] < mProx.getMaximumRange();
+            mProxIsCovered = event.values[0] < (int) mProx.getMaximumRange();
         }
 
         @Override
